requirements.txt
pytrends==4.9.2
openai==1.3.7
pandas==1.5.3
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
gspread==5.11.3
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
schedule==1.2.0
python-dateutil==2.8.2
openpyxl==3.1.2
colorlog==6.7.0
